inst_seg/requirements.txt
opencv-python~=4.6.0.66
mss~=7.0.1
timm~=0.6.11
dataclasses
ftfy
regex
fasttext
scikit-learn
lvis~=0.5.3
nltk~=3.7
git+https://github.com/openai/CLIP.git

detectron2~=0.6
fvcore~=0.1.5.post20220512
numpy~=1.23.3
pillow~=9.2.0
pycocotools~=2.0.5
tabulate~=0.9.0
tqdm~=4.64.1
matplotlib~=3.6.1
setuptools~=63.4.1
scipy~=1.9.3
scikit-image
pandas
matplotlib
gdown
openpyxl
open3d